Using _id and some neglected Models added

+++ src/main/kotlin/cut/the/crap/common/Location.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class Location(
     val type: String = "Point",
-    val location: List<Long> = emptyList()
+    val location: List<Double> = emptyList()
 )
